Lots of changes and improvements: Pubsub API: * Add a module_poisonpill() function, to send a system message to a module to stop it -> this will be enqueued in module's message queue, conversely to module_stop that should stop module right away freeing all its enqueued messages * Always destroy messages in flush_pubsub_msg(), when a module gets stopped or a loop stops looping
Map API:
* Fixed value update in map implementation and added a specific test
* Improved hashmap implementation, switching to https://github.com/DavidLeeds/hashmap as base
* Added a stress test for hashmap, with 1mln of key,value pairs
* Added map_iterate tests

Stack API:
* Stack_iterate() follows same logic as map_iterate

Generic:
* Renamed modules_set_memalloc_hook to modules_set_memhook()
* Renamed memalloc_hook to memhook_t

+++ Lib/public/module/stack.h
@@ -39,7 +39,7 @@ _public_ void *stack_pop(stack_t *s);
 _public_ void *stack_peek(const stack_t *s);
 _public_ stack_ret_code stack_clear(stack_t *s);
 _public_ stack_ret_code stack_free(stack_t *s);
-_public_ int stack_length(const stack_t *s);
+_public_ ssize_t stack_length(const stack_t *s);
 _public_ stack_ret_code stack_set_dtor(stack_t *s, stack_dtor fn);
 
 #ifdef __cplusplus
